+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2021 - 2021/03/01
+- Wrap dependency checking, security analysis and complexity/ Maintainability
+  checking in 'audit'
+- Tidy up
+- Sorted out some typing
+
 ## 2020.5.1 - 2020/10/17
 - Bump simplesecurity version